[Refactor] refactoring for mTSP
