+++ rl4co/envs/routing/mtsp/render.py
@@ -1,0 +1,95 @@
+import torch
+import numpy as np
+import matplotlib.pyplot as plt
+
+from matplotlib import colormaps
+
+from rl4co.utils.pylogger import get_pylogger
+
+log = get_pylogger(__name__)
+
+
+def render(td, actions=None, ax=None):
+    def discrete_cmap(num, base_cmap="nipy_spectral"):
+        """Create an N-bin discrete colormap from the specified input map"""
+        base = colormaps[base_cmap]
+        color_list = base(np.linspace(0, 1, num))
+        cmap_name = base.name + str(num)
+        return base.from_list(cmap_name, color_list, num)
+
+    if actions is None:
+        actions = td.get("action", None)
+    # if batch_size greater than 0 , we need to select the first batch element
+    if td.batch_size != torch.Size([]):
+        td = td[0]
+        actions = actions[0]
+
+    num_agents = td["num_agents"]
+    locs = td["locs"]
+    cmap = discrete_cmap(num_agents, "rainbow")
+
+    fig, ax = plt.subplots()
+
+    # Add depot action = 0 to before first action and after last action
+    actions = torch.cat(
+        [
+            torch.zeros(1, dtype=torch.int64),
+            actions,
+            torch.zeros(1, dtype=torch.int64),
+        ]
+    )
+
+    # Make list of colors from matplotlib
+    for i, loc in enumerate(locs):
+        if i == 0:
+            # depot
+            marker = "s"
+            color = "g"
+            label = "Depot"
+            markersize = 10
+        else:
+            # normal location
+            marker = "o"
+            color = "tab:blue"
+            label = "Cities"
+            markersize = 8
+        if i > 1:
+            label = ""
+
+        ax.plot(
+            loc[0],
+            loc[1],
+            color=color,
+            marker=marker,
+            markersize=markersize,
+            label=label,
+        )
+
+    # Plot the actions in order
+    agent_idx = 0
+    for i in range(len(actions)):
+        if actions[i] == 0:
+            agent_idx += 1
+        color = cmap(num_agents - agent_idx)
+
+        from_node = actions[i]
+        to_node = (
+            actions[i + 1] if i < len(actions) - 1 else actions[0]
+        )  # last goes back to depot
+        from_loc = td["locs"][from_node]
+        to_loc = td["locs"][to_node]
+        ax.plot([from_loc[0], to_loc[0]], [from_loc[1], to_loc[1]], color=color)
+        ax.annotate(
+            "",
+            xy=(to_loc[0], to_loc[1]),
+            xytext=(from_loc[0], from_loc[1]),
+            arrowprops=dict(arrowstyle="->", color=color),
+            annotation_clip=False,
+        )
+
+    # Legend
+    handles, labels = ax.get_legend_handles_labels()
+    ax.legend(handles, labels)
+    ax.set_title("mTSP")
+    ax.set_xlabel("x-coordinate")
+    ax.set_ylabel("y-coordinate")
